stateless validation refactor
