+++ irohad/validation/CMakeLists.txt
@@ -1,4 +1,3 @@
-add_subdirectory(stateless)
 add_subdirectory(chain)
 
 add_library(stateful_validator
@@ -9,5 +8,14 @@ target_link_libraries(stateful_validator
     optional
     ed25519
     rxcpp
+    model
+    )
+
+
+add_library(stateless_validator
+    impl/stateless_validator_impl.cpp
+    )
+
+target_link_libraries(stateless_validator
     model
     )
