+++ irohad/validation/impl/stateless_validator_impl.hpp
@@ -15,25 +15,28 @@
  * limitations under the License.
  */
 
-#ifndef IROHA_TRANSACTION_VALIDATOR_IMPL_HPP
-#define IROHA_TRANSACTION_VALIDATOR_IMPL_HPP
+#ifndef IROHA_STATELESS_VALIDATOR_IMPL_HPP
+#define IROHA_STATELESS_VALIDATOR_IMPL_HPP
 
-#include <validation/stateless/validator.hpp>
+#include "model/model_crypto_provider.hpp"
+#include "validation/stateless_validator.hpp"
 
 namespace iroha {
   namespace validation {
+
     class StatelessValidatorImpl : public StatelessValidator {
      public:
       explicit StatelessValidatorImpl(
           model::ModelCryptoProvider& crypto_provider);
       bool validate(const model::Transaction& transaction) const override;
+      bool validate(const model::Query& query) const override;
 
      private:
       static constexpr uint64_t MAX_DELAY =
           1000 * 3600 * 24;  // max-delay between tx creation and validation
       const model::ModelCryptoProvider& crypto_provider_;
     };
-  }  // namespace validation
-}  // namespace iroha
+  }
+}
 
-#endif  // IROHA_TRANSACTION_VALIDATOR_IMPL_HPP
+#endif  // IROHA_STATELESS_VALIDATOR_IMPL_HPP
