--- irohad/validation/stateless/CMakeLists.txt
@@ -1,11 +1,0 @@
-add_library(stateless_validator
-    validator_impl.cpp
-    )
-
-target_link_libraries(stateless_validator
-    schema
-    optional
-    ed25519
-    rxcpp
-    model
-    )
