--- irohad/validation/stateless/validator.hpp
@@ -1,41 +1,0 @@
-/*
-Copyright Soramitsu Co., Ltd. 2016 All Rights Reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-     http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-#ifndef IROHA_VALIDATION_STATELESS_VALIDATOR_HPP
-#define IROHA_VALIDATION_STATELESS_VALIDATOR_HPP
-
-#include <model/transaction.hpp>
-
-namespace iroha {
-  namespace validation {
-
-    /**
-     * Interface for performing stateless validation
-     */
-    class StatelessValidator {
-     public:
-      /**
-       * Perform stateless validation on a given transaction
-       * @param transaction transaction to be validated
-       * @return true if given transaction is valid, false otherwise
-       */
-      virtual bool validate(const model::Transaction &transaction) const = 0;
-
-      virtual ~StatelessValidator() = default;
-    };
-  }  // namespace validation
-}  // namespace iroha
-
-#endif  // IROHA_VALIDATION_STATELESS_VALIDATOR_HPP
