+++ irohad/validation/stateless_validator.hpp
@@ -1,0 +1,32 @@
+/**
+ * Copyright Soramitsu Co., Ltd. 2017 All Rights Reserved.
+ * http://soramitsu.co.jp
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef IROHA_STATELESS_VALIDATOR_HPP
+#define IROHA_STATELESS_VALIDATOR_HPP
+
+namespace iroha {
+  namespace validation {
+
+    class StatelessValidator {
+     public:
+      virtual bool validate(const model::Transaction& transaction) const = 0;
+      virtual bool validate(const model::Query& query) const = 0;
+    };
+  }
+}
+
+#endif  // IROHA_STATELESS_VALIDATOR_HPP
