+++ test/module/irohad/validation/stateless/transaction_validator_test.cpp
@@ -19,7 +19,7 @@
 #include <crypto/crypto.hpp>
 #include <model/model_crypto_provider_impl.hpp>
 #include <model/model_hash_provider_impl.hpp>
-#include <validation/stateless/validator_impl.hpp>
+#include "validation/impl/stateless_validator_impl.hpp"
 #include <chrono>
 
 using namespace iroha::model;
